+++ arch/arm/src/armv7-a/arm_gicv2.c
@@ -30,6 +30,7 @@
 #include <errno.h>
 
 #include <nuttx/arch.h>
+#include <nuttx/pci/pci.h>
 #include <nuttx/spinlock.h>
 #include <arch/irq.h>
 
@@ -775,4 +776,28 @@ void up_send_smp_call(cpu_set_t cpuset)
   up_trigger_irq(GIC_SMP_CPUCALL, cpuset);
 }
 #endif
+
+/****************************************************************************
+ * Name: up_get_legacy_irq
+ *
+ * Description:
+ *   Reserve vector for legacy
+ *
+ ****************************************************************************/
+
+int up_get_legacy_irq(uint32_t devfn, uint8_t line, uint8_t pin)
+{
+#if CONFIG_ARMV7A_GICV2_LEGACY_IRQ0 >= 0
+  uint8_t slot;
+  uint8_t tmp;
+
+  UNUSED(line);
+  slot = PCI_SLOT(devfn);
+  tmp = (pin - 1 + slot) % 4;
+  return CONFIG_ARMV7A_GICV2_LEGACY_IRQ0 + tmp;
+#else
+  return -ENOTSUP;
+#endif
+}
+
 #endif /* CONFIG_ARMV7A_HAVE_GICv2 */
